+++ tests/unit/task/contexts/designate/test_zones.py
@@ -67,6 +67,7 @@ class ZoneGeneratorTestCase(test.ScenarioTestCase):
                 },
                 "zones": {
                     "zones_per_tenant": zones_per_tenant,
+                    "set_zone_in_network": False
                 }
             },
             "admin": {
@@ -85,6 +86,55 @@ class ZoneGeneratorTestCase(test.ScenarioTestCase):
         zones_ctx = zones.ZoneGenerator(self.context)
         zones_ctx.setup()
         self.assertEqual(new_context, self.context)
+
+    @mock.patch("%s.neutron.utils.NeutronScenario" % SCN)
+    @mock.patch("%s.designate.utils.DesignateScenario._create_zone" % SCN,
+                return_value={"id": "uuid", "name": "fake_name"})
+    def test_setup_for_existinge(self, mock_designate_scenario__create_zone,
+                                 mock_neutron_scenario):
+        tenants_count = 1
+        users_per_tenant = 1
+
+        networks = []
+        tenants = self._gen_tenants(tenants_count)
+        users = []
+        for id_ in tenants.keys():
+            networks.append(
+                {"id": f"foo_net_{id_}",
+                 "tenant_id": id_, "subnets": ["foo_subnet"]})
+            for i in range(users_per_tenant):
+                users.append({"id": i, "tenant_id": id_,
+                              "credential": mock.MagicMock()})
+        tenants["0"]["networks"] = networks
+
+        self.context.update({
+            "config": {
+                "users": {
+                    "tenants": 1,
+                    "users_per_tenant": 1,
+                    "concurrent": 1,
+                },
+                "zones": {
+                    "set_zone_in_network": True
+                },
+                "network": {}
+            },
+            "admin": {
+                "credential": mock.MagicMock()
+            },
+            "users": users,
+            "tenants": tenants
+        })
+
+        zones_ctx = zones.ZoneGenerator(self.context)
+        zones_ctx.setup()
+
+        mock_neutron_scenario.assert_called_once()
+        scenario = mock_neutron_scenario.return_value
+        scenario.clients.assert_called_with("neutron")
+        neutron = scenario.clients.return_value
+        neutron.update_network.assert_called_with(
+            "foo_net_0", {"network": {"dns_domain": "fake_name"}})
 
     @mock.patch("%s.designate.zones.resource_manager.cleanup" % CTX)
     def test_cleanup(self, mock_cleanup):